minor code changes
